pyproject.toml
[tool.poetry]
name = "fullon-master-api"
version = "1.0.0"
description = "Unified API Gateway for Fullon Trading Platform"
authors = ["Fullon Team"]
readme = "README.md"
packages = [{include = "fullon_master_api", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.0"
websockets = "^12.0"

# Fullon core libraries (REQUIRED - used directly)
fullon-log = {git = "ssh://git@github.com/ingmarAvocado/fullon_log.git", branch = "main"}
fullon-orm = {git = "ssh://git@github.com/ingmarAvocado/fullon_orm.git", branch = "main"}
fullon-cache = {git = "ssh://git@github.com/ingmarAvocado/fullon_cache.git", branch = "main"}
fullon-ohlcv = {git = "ssh://git@github.com/ingmarAvocado/fullon_ohlcv.git", branch = "main"}
fullon-credentials = {git = "ssh://git@github.com/ingmarAvocado/fullon_credentials.git", branch = "main"}
fullon-exchange = {git = "ssh://git@github.com/ingmarAvocado/fullon_exchange.git", branch = "main"}

# Fullon APIs (router composition)
fullon-orm-api = {path = "../fullon_orm_api", develop = true}
fullon-ohlcv-api = {path = "../fullon_ohlcv_api", develop = true}
fullon-cache-api = {path = "../fullon_cache_api", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
httpx = "^0.25.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.6.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
